# backend/api/openapi.yaml
openapi: 3.1.0
info:
  title: StreamPulse API
  version: 1.0.0
  summary: Real-time audio streaming platform.
  description: |
    REST + SSE API of the StreamPulse platform (Go 1.26, chi router).

    ## Response envelope

    Every JSON response produced by a handler is wrapped:

    ```json
    { "data": { }, "meta": { "requestId": "uuid", "timestamp": "RFC3339" } }
    ```

    List endpoints add `page`, `perPage` and `total` to `meta`.
    Errors replace `data` with `error: { code, message }`.

    ## Correlating a response with the logs

    Every response carries an `X-Request-Id` header, and `meta.requestId`
    repeats it. The same value is emitted as `request_id` on the server's
    structured log line for that request, alongside the OpenTelemetry
    `trace_id`. Quote it in a bug report and the request can be found in the
    logs and in the trace.

    The header is the more reliable of the two: it is present even on the
    middleware errors described below, which have no `meta` envelope.

    ## Roles

    Roles are ordered: `anonymous` < `user` < `broadcaster` < `admin`.
    An endpoint marked "broadcaster" is therefore also reachable by an admin.

    ## Personal data

    `GET /users/me` returns every piece of personal data held on the
    calling account; `PATCH /users/me` rectifies its email or username;
    `DELETE /users/me` erases the account and everything attached to it
    (tokens, streams, playlists, favorites, uploaded tracks). An admin can
...

...

...
servers:
...
security: []
paths:
...
  /auth/oauth:
    post:
      tags: [auth]
      operationId: oauthLogin
      summary: Open a session with Google or Apple
      description: |
        Social login. The mobile app obtains an ID token from the provider's
        SDK (Google Sign-In or Sign in with Apple) and posts it here. The
        server verifies its signature against the provider's JWKS, its issuer
        and its audience (`GOOGLE_OAUTH_CLIENT_IDS` / `APPLE_OAUTH_CLIENT_IDS`),
        then opens a session. First sign-in creates a `user` account, or links
        the identity to the existing account bearing the same verified email.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OAuthLoginRequest'
      responses:
        '200':
          description: Session opened (account created on first sign-in).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenEnvelope'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401':
          description: Identity token rejected (bad signature, issuer, audience, or expired).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
              example:
                error: { code: UNAUTHORIZED, message: invalid identity token }
                meta: { requestId: 8f1c…, timestamp: '2026-08-27T10:00:00Z' }
        '409':
          description: Email already registered but not verified by the provider, so it cannot be linked.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
              example:
                error: { code: CONFLICT, message: email already registered with a password account }
                meta: { requestId: 8f1c…, timestamp: '2026-08-27T10:00:00Z' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '500': { $ref: '#/components/responses/InternalError' }
        '503':
          description: The requested provider has no client ID configured on this server.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
              example:
                error: { code: SERVICE_UNAVAILABLE, message: social login is not configured on this server }
                meta: { requestId: 8f1c…, timestamp: '2026-08-27T10:00:00Z' }
  /auth/refresh:
    post:
      tags: [auth]
      operationId: refreshToken
      summary: Rotate the token pair
      description: |
        Exchanges a refresh token for a fresh pair. The presented refresh
        token is consumed: rotation is one-shot, replaying it returns 401.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: New token pair.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenEnvelope'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401':
          description: Refresh token unknown, already used, or expired.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
              example:
                error: { code: UNAUTHORIZED, message: invalid or expired refresh token }
                meta: { requestId: 8f1c…, timestamp: '2026-08-27T10:00:00Z' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '500': { $ref: '#/components/responses/InternalError' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        `Authorization: Bearer <access_token>`, HS256, 15 minutes by default
        (`JWT_EXPIRY`). Claims carry `user_id`, `username` and `role`.
        Renew with `POST /auth/refresh`.
  headers:
    RequestID:
      description: |
        Identifier of this request, generated by the server. Repeated in
        `meta.requestId` when the response has an envelope, and emitted as
        `request_id` on the matching structured log line.
      schema:
        type: string
        example: streampulse/8f1c4a2b0e91/000042
...
  responses:
    BadRequest:
      description: Malformed body. Unknown JSON fields are rejected.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestID'
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
          example:
            error: { code: BAD_REQUEST, message: invalid request body }
            meta: { requestId: 8f1c…, timestamp: '2026-08-27T10:00:00Z' }
...
    InternalError:
      description: Unexpected server-side failure.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestID'
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
          example:
            error: { code: INTERNAL_ERROR, message: failed to list streams }
            meta: { requestId: 8f1c…, timestamp: '2026-08-27T10:00:00Z' }
...
    RateLimited:
      description: |
        Per-IP token bucket exhausted (`RATE_LIMIT_RPS`, burst
        `RATE_LIMIT_BURST`). Written by the rate-limit middleware: bare
        `error` object, `text/plain` content type.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestID'
      content:
        text/plain:
          schema: { $ref: '#/components/schemas/BareError' }
          example:
            error: { code: RATE_LIMITED, message: too many requests }
  schemas:
    Meta:
      type: object
      required: [requestId, timestamp]
      properties:
        requestId:
          type: string
          description: |
            Identifier of the request, identical to the `X-Request-Id`
            response header and to the `request_id` field of the matching
            structured log line. Not a UUID: it is the chi request id, of the
            form `<host>/<random>-<counter>`.
        timestamp:
          type: string
          format: date-time
          description: UTC, RFC 3339.
        page:
          type: integer
          description: Present on list endpoints only.
        perPage:
          type: integer
          description: Present on list endpoints only.
        total:
          type: integer
          description: |
            Total row count, present on list endpoints only. Serialized with
            `omitempty`, so it is **absent** rather than `0` when the result
            set is empty — treat a missing `total` as zero.
    Envelope:
      type: object
      required: [data, meta]
      properties:
        data: true
        meta: { $ref: '#/components/schemas/Meta' }
...
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Stable, machine-readable code.
          enum:
            - BAD_REQUEST
            - INVALID_ID
            - INVALID_BODY
            - INVALID_INPUT
            - UNAUTHORIZED
            - FORBIDDEN
            - NOT_FOUND
            - CONFLICT
            - STREAM_NOT_LIVE
            - RATE_LIMITED
            - INTERNAL_ERROR
        message:
          type: string
          description: Human-readable, English, not meant to be shown as-is to end users.
    ErrorEnvelope:
      type: object
      required: [error, meta]
      properties:
        error: { $ref: '#/components/schemas/Error' }
        meta: { $ref: '#/components/schemas/Meta' }
    BareError:
      type: object
      description: Error shape emitted by the middleware chain, without the `meta` envelope.
      required: [error]
      properties:
        error: { $ref: '#/components/schemas/Error' }
...
    Role:
      type: string
      description: Ordered from least to most privileged.
      enum: [anonymous, user, broadcaster, admin]
...
    User:
      type: object
      required: [id, email, username, role]
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        username: { type: string }
        role: { $ref: '#/components/schemas/Role' }
...
    TokenPair:
      type: object
      required: [access_token, refresh_token, expires_at, user]
      properties:
        access_token:
          type: string
          description: 'JWT HS256. Send it as the `Authorization: Bearer …` header.'
        refresh_token:
          type: string
          description: Opaque, single use, 168 h by default (`JWT_REFRESH_EXPIRY`).
        expires_at:
          type: string
          format: date-time
          description: Expiry of the access token, not of the refresh token.
        user: { $ref: '#/components/schemas/User' }
...
    RefreshRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token: { type: string }
    OAuthLoginRequest:
      type: object
      required: [provider, id_token]
      properties:
        provider:
          type: string
          enum: [google, apple]
        id_token:
          type: string
          description: Signed ID token returned by the provider's SDK on the device.
      example:
        provider: google
        id_token: eyJhbGciOiJSUzI1NiIs…
...
    TokenEnvelope:
      allOf:
        - $ref: '#/components/schemas/Envelope'
        - type: object
          properties:
            data: { $ref: '#/components/schemas/TokenPair' }
...