# backend/internal/transport/http/dto/auth_dto.go
package dto

import "time"

type RegisterRequest struct {
	Email    string `json:"email"`
	Username string `json:"username"`
	Password string `json:"password"`
	// AcceptedTerms doit valoir true : case a cocher obligatoire de
	// l'ecran d'inscription, revalidee ici pour ne pas dependre du seul
	// client (docs/rgpd.md).
	AcceptedTerms bool `json:"accepted_terms"`
}

type LoginRequest struct {
	Email    string `json:"email"`
	Password string `json:"password"`
}

type RefreshRequest struct {
	RefreshToken string `json:"refresh_token"`
}

type OAuthLoginRequest struct {
	// Provider : "google" ou "apple".
	Provider string `json:"provider"`
	// IDToken : l'ID token signe rendu par le SDK du fournisseur sur
	// l'appareil ; le serveur le verifie contre le JWKS du fournisseur.
	IDToken string `json:"id_token"`
}

type TokenResponse struct {
	AccessToken  string    `json:"access_token"`
	RefreshToken string    `json:"refresh_token"`
	ExpiresAt    time.Time `json:"expires_at"`
	User         UserDTO   `json:"user"`
}

type UserDTO struct {
	ID       string `json:"id"`
	Email    string `json:"email"`
	Username string `json:"username"`
	Role     string `json:"role"`
}